refactor: pickup 관련 엔티티들 pickup 패키지로 이동
