+++ BE/src/main/java/ifive/idrop/driver/domain/Driver.java
@@ -4,7 +4,7 @@ import ifive.idrop.driver.dto.DriverInformation;
 import ifive.idrop.driver.dto.DriverDetailResponse;
 import ifive.idrop.driver.dto.DriverSummary;
 import ifive.idrop.driver.dto.WorkHoursDto;
-import ifive.idrop.entity.PickUpSubscription;
+import ifive.idrop.pickup.domain.PickUpSubscription;
 import ifive.idrop.user.domain.User;
 import ifive.idrop.common.enums.Gender;
 import ifive.idrop.common.exception.BusinessException;
