+++ BE/src/main/java/ifive/idrop/driver/dto/DriverSubscribeInfoResponse.java
@@ -1,8 +1,9 @@
 package ifive.idrop.driver.dto;
 
 import ifive.idrop.child.domain.Child;
-import ifive.idrop.entity.*;
 import ifive.idrop.parent.domain.Parent;
+import ifive.idrop.pickup.domain.PickUpLocation;
+import ifive.idrop.pickup.domain.PickUpSubscription;
 import lombok.Builder;
 import lombok.Getter;
 import org.json.simple.JSONObject;
