+++ BE/src/main/java/ifive/idrop/dto/CurrentPickUpResponse.java
@@ -1,8 +1,8 @@
 package ifive.idrop.dto;
 
 import com.fasterxml.jackson.annotation.JsonUnwrapped;
-import ifive.idrop.entity.PickUpSubscription;
-import ifive.idrop.entity.PickUpLocation;
+import ifive.idrop.pickup.domain.PickUpSubscription;
+import ifive.idrop.pickup.domain.PickUpLocation;
 import lombok.*;
 
 import java.time.LocalDateTime;
