+++ BE/src/main/java/ifive/idrop/parent/service/ParentService.java
@@ -10,26 +10,22 @@ import ifive.idrop.dto.CurrentPickUpResponse;
 
 import ifive.idrop.parent.dto.PickUpHistoryResponse;
 import ifive.idrop.parent.dto.ParentSubscribeInfoResponse;
-import ifive.idrop.entity.*;
-import ifive.idrop.entity.enums.PickUpStatus;
+import ifive.idrop.pickup.domain.*;
+import ifive.idrop.pickup.domain.enums.PickUpStatus;
 import ifive.idrop.common.exception.BusinessException;
 import ifive.idrop.common.exception.ErrorCode;
-import ifive.idrop.notification.AlarmMessage;
-import ifive.idrop.notification.NotificationUtill;
 import ifive.idrop.driver.repository.DriverRepository;
 import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.parent.repository.ParentRepository;
-import ifive.idrop.pickup.domain.PickUp;
 import ifive.idrop.pickup.repository.PickUpRepository;
-import jakarta.persistence.criteria.CriteriaBuilder;
 import lombok.RequiredArgsConstructor;
 import org.json.JSONException;
-import org.json.simple.JSONObject;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.time.LocalTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
@@ -44,29 +40,19 @@ public class ParentService {
 
     @Transactional
     public BaseResponse<String> subscribe(Parent parent, SubscribeRequest subscribeRequest) throws JSONException, ExecutionException, InterruptedException {
-        Map<Day, Map<String, Integer>> schedule = subscribeRequest.getSchedule();
-        for (Map.Entry<Day, Map<String, Integer>> entry : schedule.entrySet()) {
-            Day day = entry.getKey();
-            int hour = entry.getValue().get("hour");
-            int min = entry.getValue().get("min");
-            System.out.println(day + " " + hour + " " + min);
-        }
-        return BaseResponse.success();
 
+        Driver driver = driverRepository.findById(subscribeRequest.getDriverId())
+                .orElseThrow(() -> new BusinessException(ErrorCode.DRIVER_NOT_EXIST));
+        Child child = parentRepository.findChild(parent.getId())
+                .orElseThrow(() -> new BusinessException(ErrorCode.CHILD_NOT_EXIST));
 
-//
-//        Driver driver = driverRepository.findById(subscribeRequest.getDriverId())
-//                .orElseThrow(() -> new BusinessException(ErrorCode.DRIVER_NOT_EXIST));
-//        Child child = parentRepository.findChild(parent.getId())
-//                .orElseThrow(() -> new BusinessException(ErrorCode.CHILD_NOT_EXIST));
-//
-//        PickUpLocation location = createPickUpLocation(subscribeRequest);
-//        createPickUpInfo(subscribeRequest, child, driver, location);
-//
-//        // 요청한 기사에게 알람
+        PickUpLocation location = createPickUpLocation(subscribeRequest);
+        createPickUpInfo(subscribeRequest, child, driver, location);
+
+        // 요청한 기사에게 알람
 //        NotificationUtill.createNotification(driver, AlarmMessage.SUBSCRIBE_REQUEST.getTitle(),
 //                AlarmMessage.SUBSCRIBE_REQUEST.getMessage());
-//        return BaseResponse.success();
+        return BaseResponse.success();
     }
 
     public BaseResponse<List<CurrentPickUpResponse>> getChildRunningInfo(Parent parent) {
@@ -83,22 +69,18 @@ public class ParentService {
                 .driver(driver)
                 .status(PickUpStatus.WAIT)
                 .requestDate(LocalDateTime.now())
+                .pickUpScheduleList(new ArrayList<>())
                 .build();
         pickUpSubscription.updatePickUpLocation(location);
         pickUpRepository.savePickUpInfo(pickUpSubscription);
-        pickUpSubscription.getId();
 
-//        Map<String, Map<Day, Integer>> schedule = subscribeRequest.getSchedule();
-//        for (Map.Entry<String, Map<String, Integer>> entry : schedule.entrySet()) {
-//            String day = entry.getKey();
-//            int hour = entry.getValue().get("hour");
-//            int min = entry.getValue().get("min");
-//            pickUpSubscription.addPickUpSchedule(new PickUpSchedule(new PickUpScheduleId(pickUpSubscription.getId(), day), LocalTime.of(hour, min)));
-////            System.out.println(key);
-////            for (Map.Entry<String, Integer> a : obj.get(key).entrySet()) {
-////                System.out.println(a.getKey() + " " +  a.getValue());
-////            }
-//        }
+        Map<Day, Map<String, Integer>> schedule = subscribeRequest.getSchedule();
+        for (Map.Entry<Day, Map<String, Integer>> entry : schedule.entrySet()) {
+            Day day = entry.getKey();
+            int hour = entry.getValue().get("hour");
+            int min = entry.getValue().get("min");
+            pickUpSubscription.addPickUpSchedule(new PickUpSchedule(new PickUpScheduleId(pickUpSubscription.getId(), day), LocalTime.of(hour, min), pickUpSubscription));
+        }
         return pickUpSubscription;
     }
