+++ BE/src/main/java/ifive/idrop/pickup/domain/PickUpScheduleId.java
@@ -1,4 +1,4 @@
-package ifive.idrop.entity;
+package ifive.idrop.pickup.domain;
 
 import ifive.idrop.common.enums.Day;
 import jakarta.persistence.Embeddable;
