+++ BE/src/main/java/ifive/idrop/websocket/WebSocketHandShakeInterceptor.java
@@ -6,7 +6,6 @@ import ifive.idrop.auth.domain.AuthenticateUser;
 import ifive.idrop.auth.filter.VerifyUserFilter;
 import ifive.idrop.auth.utils.JwtProvider;
 import ifive.idrop.user.repository.UserRepository;
-import ifive.idrop.util.CustomObjectMapper;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
