+++ BE/src/main/java/ifive/idrop/websocket/location/LocationWebSocketHandler.java
@@ -2,7 +2,6 @@ package ifive.idrop.websocket.location;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import ifive.idrop.driver.domain.Driver;
-import ifive.idrop.entity.*;
 import ifive.idrop.common.exception.BusinessException;
 import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.auth.domain.AuthenticateUser;
@@ -10,9 +9,10 @@ import ifive.idrop.auth.filter.VerifyUserFilter;
 import ifive.idrop.auth.utils.JwtProvider;
 import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.pickup.domain.PickUp;
+import ifive.idrop.pickup.domain.PickUpLocation;
 import ifive.idrop.user.repository.UserRepository;
 import ifive.idrop.user.domain.User;
-import ifive.idrop.util.CustomObjectMapper;
+import ifive.idrop.websocket.CustomObjectMapper;
 import ifive.idrop.websocket.PickUpInfoRepository;
 import ifive.idrop.websocket.location.dto.ChildGeoLocation;
 import ifive.idrop.websocket.location.dto.CurrentPickUp;
